feat: add an api to get stats info and set type of count to long

+++ src/main/java/org/fansuregrin/entity/PageResult.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 public class PageResult<T> {
-    Integer total;
+    Long total;
 
     List<T> items;
 }
